[TASK] DPL-160: Split Extension Client into functional clients
For a better overview and encapsulation of functions according to the
DeepL API, the internal `Client`, used as wrapper to the official API,
is split into dedicated functional parts.

The respective Interfaces for `Usage` and `Translator` were introduced
times ago, and now this will take effect by defining clients for their
use cases.

The Services needing the CLient were adopted to the new interfaces to
keep functionality `as-is`.

As the `Client` was set to final and therefore was not extendable, this
change should have no unwanted side effect to other extensions building
up on `deepltranslate_core`.

This was needed to split out the Glossary functionality completely from
the `deepltranslate_core` and only define the Interfaces for Glossary
API v2 and v3.

The functionality of glossary management is moved into
`deepltranslate_glossary` as it's unused inside `deepltracnslate_core`
and a dedicated functionality should be handled inside the respective
extension.
