--- Classes/Client.php
@@ -1,252 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace WebVision\Deepltranslate\Core;
-
-use DeepL\DeepLException;
-use DeepL\GlossaryEntries;
-use DeepL\GlossaryInfo;
-use DeepL\GlossaryLanguagePair;
-use DeepL\Language;
-use DeepL\TextResult;
-use DeepL\TranslateTextOptions;
-use DeepL\Usage;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\DependencyInjection\Attribute\AsAlias;
-use WebVision\Deepltranslate\Core\Client\DeepLClientInterface;
-use WebVision\Deepltranslate\Core\Exception\ApiKeyNotSetException;
-
-/**
- * @internal No public usage
- * @todo split the client into two separate services?
- */
-#[AsAlias(id: ClientInterface::class, public: true)]
-final class Client implements TranslatorInterface, UsageInterface
-{
-    //
-    public function __construct(
-        private readonly LoggerInterface $logger,
-        private readonly DeepLClientInterface $translator
-    ) {}
-
-    /**
-     * @return TextResult|TextResult[]|null
-     *
-     * @throws ApiKeyNotSetException
-     */
-    public function translate(
-        string $content,
-        ?string $sourceLang,
-        string $targetLang,
-        string $glossary = '',
-        string $formality = ''
-    ): array|TextResult|null {
-        $options = [
-            // @todo Make this configurable, either as global setting or dependency injection (factory?) / event
-            TranslateTextOptions::FORMALITY => $formality ?: 'default',
-            // @todo Make this configurable, either as global setting or dependency injection (factory?) / event
-            TranslateTextOptions::TAG_HANDLING => 'html',
-            // @todo Make this configurable, either as global setting or dependency injection (factory?) / event
-            TranslateTextOptions::TAG_HANDLING_VERSION => 'v2',
-        ];
-
-        if (!empty($glossary)) {
-            $options[TranslateTextOptions::GLOSSARY] = $glossary;
-        }
-
-        try {
-            return $this->translator->translateText(
-                $content,
-                $sourceLang,
-                $targetLang,
-                $options
-            );
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-
-        return null;
-    }
-
-    /**
-     * @return Language[]
-     *
-     * @throws ApiKeyNotSetException
-     */
-    public function getSupportedLanguageByType(string $type = 'target'): array
-    {
-        try {
-            return ($type === 'target')
-                ? $this->translator->getTargetLanguages()
-                : $this->translator->getSourceLanguages();
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-
-        return [];
-    }
-
-    /**
-     * @return GlossaryLanguagePair[]
-     *
-     * @throws ApiKeyNotSetException
-     */
-    public function getGlossaryLanguagePairs(): array
-    {
-        try {
-            return $this->translator->getGlossaryLanguages();
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-
-        return [];
-    }
-
-    /**
-     * @return GlossaryInfo[]
-     *
-     * @throws ApiKeyNotSetException
-     */
-    public function getAllGlossaries(): array
-    {
-        try {
-            return $this->translator->listGlossaries();
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-
-        return [];
-    }
-
-    /**
-     * @throws ApiKeyNotSetException
-     */
-    public function getGlossary(string $glossaryId): ?GlossaryInfo
-    {
-        try {
-            return $this->translator->getGlossary($glossaryId);
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-
-        return null;
-    }
-
-    /**
-     * @param array<int, array{source: string, target: string}> $entries
-     *
-     * @throws ApiKeyNotSetException
-     */
-    public function createGlossary(
-        string $glossaryName,
-        string $sourceLang,
-        string $targetLang,
-        array $entries
-    ): GlossaryInfo {
-        $prepareEntriesForGlossary = [];
-        foreach ($entries as $entry) {
-            /*
-             * as the version without trimming in TCA is already published,
-             * we trim a second time here
-             * to avoid errors in DeepL client
-             */
-            $source = trim($entry['source']);
-            $target = trim($entry['target']);
-            if (empty($source) || empty($target)) {
-                continue;
-            }
-            $prepareEntriesForGlossary[$source] = $target;
-        }
-        try {
-            return $this->translator->createGlossary(
-                $glossaryName,
-                $sourceLang,
-                $targetLang,
-                GlossaryEntries::fromEntries($prepareEntriesForGlossary)
-            );
-        } catch (DeepLException $e) {
-            return new GlossaryInfo(
-                '',
-                '',
-                false,
-                '',
-                '',
-                new \DateTime(),
-                0
-            );
-        }
-    }
-
-    /**
-     * @throws ApiKeyNotSetException
-     */
-    public function deleteGlossary(string $glossaryId): void
-    {
-        try {
-            $this->translator->deleteGlossary($glossaryId);
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-    }
-
-    /**
-     * @throws ApiKeyNotSetException
-     */
-    public function getGlossaryEntries(string $glossaryId): ?GlossaryEntries
-    {
-        try {
-            return $this->translator->getGlossaryEntries($glossaryId);
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-
-        return null;
-    }
-
-    /**
-     * @throws ApiKeyNotSetException
-     */
-    public function getUsage(): ?Usage
-    {
-        try {
-            return $this->translator->getUsage();
-        } catch (DeepLException $exception) {
-            $this->logger->error(sprintf(
-                '%s (%d)',
-                $exception->getMessage(),
-                $exception->getCode()
-            ));
-        }
-
-        return null;
-    }
-}
