+++ Classes/ClientInterface.php
@@ -4,63 +4,19 @@ declare(strict_types=1);
 
 namespace WebVision\Deepltranslate\Core;
 
-use DeepL\GlossaryEntries;
-use DeepL\GlossaryInfo;
-use DeepL\GlossaryLanguagePair;
+use Psr\Log\LoggerInterface;
 use WebVision\Deepltranslate\Core\Client\DeepLClientInterface;
-use WebVision\Deepltranslate\Core\Exception\ApiKeyNotSetException;
 
 /**
  * Interface for custom client implementation and which methods are expected.
  *
  * @internal use only for testing, not part of public extension API.
- * @property DeepLClientInterface $translator
+ * @property DeepLClientInterface $client
  */
 interface ClientInterface
 {
-    /**
-     * @return GlossaryLanguagePair[]
-     *
-     * @throws ApiKeyNotSetException
-     */
-    public function getGlossaryLanguagePairs(): array;
-
-    /**
-     * @return GlossaryInfo[]
-     *
-     * @throws ApiKeyNotSetException
-     * @deprecated This function is deprecated in favour of multilingual glossaries and will be removed in future versions
-     */
-    public function getAllGlossaries(): array;
-
-    /**
-     * @throws ApiKeyNotSetException
-     * @deprecated This function is deprecated in favour of multilingual glossaries and will be removed in future versions
-     */
-    public function getGlossary(string $glossaryId): ?GlossaryInfo;
-
-    /**
-     * @param array<int, array{source: string, target: string}> $entries
-     *
-     * @throws ApiKeyNotSetException
-     * @deprecated This function is deprecated in favour of multilingual glossaries and will be removed in future versions
-     */
-    public function createGlossary(
-        string $glossaryName,
-        string $sourceLang,
-        string $targetLang,
-        array $entries
-    ): GlossaryInfo;
-
-    /**
-     * @throws ApiKeyNotSetException
-     * @deprecated This function is deprecated in favour of multilingual glossaries and will be removed in future versions
-     */
-    public function deleteGlossary(string $glossaryId): void;
-
-    /**
-     * @throws ApiKeyNotSetException
-     * @deprecated This function is deprecated in favour of multilingual glossaries and will be removed in future versions
-     */
-    public function getGlossaryEntries(string $glossaryId): ?GlossaryEntries;
+    public function __construct(
+        LoggerInterface $logger,
+        DeepLClientInterface $client
+    );
 }
