+++ Classes/GlossaryV3Interface.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace WebVision\Deepltranslate\Core;
+
+interface GlossaryV3Interface extends ClientInterface {}
