+++ Classes/Service/UsageService.php
@@ -10,7 +10,6 @@ use TYPO3\CMS\Backend\Toolbar\InformationStatus;
 use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
 use TYPO3\CMS\Core\Localization\Locales;
 use TYPO3\CMS\Core\Type\ContextualFeedbackSeverity;
-use WebVision\Deepltranslate\Core\ClientInterface;
 use WebVision\Deepltranslate\Core\Event\Listener\UsageToolBarEventListener;
 use WebVision\Deepltranslate\Core\Hooks\UsageProcessAfterFinishHook;
 use WebVision\Deepltranslate\Core\UsageInterface;
@@ -21,13 +20,9 @@ use WebVision\Deepltranslate\Core\UsageInterface;
 #[Autoconfigure(public: true)]
 final readonly class UsageService implements UsageServiceInterface
 {
-    /**
-     * @param UsageInterface $client
-     * @param Locales $locales
-     */
     public function __construct(
-        private ClientInterface $client,
-        private Locales $locales,
+        private readonly UsageInterface $client,
+        private readonly Locales $locales,
     ) {}
 
     public function getCurrentUsage(): ?Usage
