+++ Tests/Functional/Client/DeepLClientTest.php
@@ -19,6 +19,7 @@ final class DeepLClientTest extends FunctionalTestCase
 
     protected array $coreExtensionsToLoad = [
         'typo3/cms-setup',
+        'typo3/cms-install',
     ];
 
     protected array $configurationToUseInTestInstance = [
